Add remaining test helpers and fix test compilation issues
Added test helpers:
- proposal::create_test_proposal() - Create real proposals for testing
- proposal::destroy_for_testing() - Clean up proposal objects
- fee::create_fake_admin_cap_for_testing() - Create test admin cap
- coin_escrow::create_test_escrow_with_market_state() - Flexible escrow creation

Fixed test issues:
- swap_position_registry::destroy_for_testing() - Use destroy_empty for table
- Type casting fixes in test files (u8 ↔ u64)
- Fixed invalid address format (@0xWRONG → @0x9999)

All tests now compile successfully.
Build succeeds with no errors.

+++ contracts/futarchy_markets_core/tests/swap_position_registry_tests.move
@@ -37,7 +37,7 @@ fun create_test_escrow<AssetType, StableType>(
     outcome_count: u8,
     ctx: &mut TxContext,
 ): TokenEscrow<AssetType, StableType> {
-    coin_escrow::create_test_escrow(outcome_count, ctx)
+    coin_escrow::create_test_escrow((outcome_count as u64), ctx)
 }
 
 /// Cleanup registry
